Move test data factory to the package level.
