+++ registration/src/test/java/org/spine3/samples/lobby/registration/order/OrderAggregateShould.java
@@ -20,12 +20,9 @@
 
 package org.spine3.samples.lobby.registration.order;
 
-import com.google.common.collect.ImmutableList;
 import com.google.protobuf.Message;
-import com.google.protobuf.util.TimeUtil;
 import org.junit.Before;
 import org.junit.Test;
-import org.spine3.base.CommandContext;
 import org.spine3.money.Money;
 import org.spine3.samples.lobby.common.ConferenceId;
 import org.spine3.samples.lobby.common.OrderId;
@@ -36,15 +33,10 @@ import java.lang.reflect.Method;
 import java.util.List;
 
 import static com.google.common.base.Throwables.propagate;
-import static com.google.common.collect.Lists.newArrayList;
 import static java.util.Collections.singletonList;
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.fail;
 import static org.spine3.money.Currency.USD;
-import static org.spine3.samples.lobby.common.util.IdFactory.newConferenceId;
-import static org.spine3.samples.lobby.common.util.IdFactory.newOrderId;
-import static org.spine3.samples.lobby.registration.testdata.TestDataFactory.newPersonalInfo;
-import static org.spine3.samples.lobby.registration.util.Seats.newSeatQuantity;
 import static org.spine3.util.MoneyUtil.newMoney;
 
 /**
@@ -318,167 +310,7 @@ public class OrderAggregateShould {
         }
     }
 
-    private static class Given {
-
-        static final OrderId ORDER_ID = newOrderId();
-        static final ConferenceId CONFERENCE_ID = newConferenceId();
-        static final List<SeatQuantity> SEATS = ImmutableList.of(newSeatQuantity(5), newSeatQuantity(10));
-
-        private final TestOrderAggregate aggregate;
-
-        private Given() {
-            aggregate = new TestOrderAggregate(ORDER_ID);
-            aggregate.setOrderPricingService(new PricingServiceStub());
-        }
-
-        TestOrderAggregate newOrder() {
-            return aggregate;
-        }
-
-        TestOrderAggregate confirmedOrder() {
-            final Order state = orderState(seats(), true);
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        TestOrderAggregate placedOrder() {
-            final Order state = orderState(seats());
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        TestOrderAggregate completelyReservedOrder(List<SeatQuantity> reservedSeats) {
-            final Order state = orderState(reservedSeats);
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        TestOrderAggregate partiallyReservedOrder(List<SeatQuantity> reservedSeats) {
-            final List<SeatQuantity> requestedSeats = newArrayList(reservedSeats);
-            //noinspection LocalVariableNamingConvention
-            final int partlyReservedSeatIndex = 0;
-            final SeatQuantity.Builder seat = requestedSeats.get(partlyReservedSeatIndex).toBuilder();
-            seat.setQuantity(seat.getQuantity() + 5);
-            requestedSeats.set(partlyReservedSeatIndex, seat.build());
-            final Order state = orderState(requestedSeats);
-            aggregate.incrementState(state);
-            return aggregate;
-        }
-
-        static List<SeatQuantity> seats() {
-            //noinspection ReturnOfCollectionOrArrayField
-            return SEATS;
-        }
-
-        static Order orderState(Iterable<SeatQuantity> seats) {
-            final Order.Builder order = Order.newBuilder()
-                    .setId(ORDER_ID)
-                    .setConferenceId(CONFERENCE_ID)
-                    .addAllSeat(ImmutableList.copyOf(seats));
-            return order.build();
-        }
-
-        static Order orderState(Iterable<SeatQuantity> seats, boolean isConfirmed) {
-            final Order.Builder order = orderState(seats).toBuilder();
-            order.setIsConfirmed(isConfirmed);
-            return order.build();
-        }
-
-        private static class Command {
-
-            static final CommandContext CMD_CONTEXT = CommandContext.getDefaultInstance();
-
-            static final RegisterToConference REGISTER_TO_CONFERENCE = RegisterToConference.newBuilder()
-                    .setOrderId(ORDER_ID)
-                    .setConferenceId(CONFERENCE_ID)
-                    .addAllSeat(SEATS)
-                    .build();
-
-            static final MarkSeatsAsReserved MARK_SEATS_AS_RESERVED = MarkSeatsAsReserved.newBuilder()
-                    .setOrderId(ORDER_ID)
-                    .setReservationExpiration(TimeUtil.getCurrentTime())
-                    .addAllSeat(SEATS)
-                    .build();
-
-            static final RejectOrder REJECT_ORDER = RejectOrder.newBuilder().setOrderId(ORDER_ID).build();
-
-            static final ConfirmOrder CONFIRM_ORDER = ConfirmOrder.newBuilder().setOrderId(ORDER_ID).build();
-
-            static final AssignRegistrantDetails ASSIGN_REGISTRANT_DETAILS = AssignRegistrantDetails.newBuilder()
-                    .setOrderId(ORDER_ID)
-                    .setRegistrant(newPersonalInfo("John", "Black", "jblack@gmail.com"))
-                    .build();
-
-            static CommandContext context() {
-                return CMD_CONTEXT;
-            }
-
-            static RegisterToConference registerToConference() {
-                return REGISTER_TO_CONFERENCE;
-            }
-
-            static MarkSeatsAsReserved markSeatsAsReserved() {
-                return MARK_SEATS_AS_RESERVED;
-            }
-
-            static ConfirmOrder confirmOrder() {
-                return CONFIRM_ORDER;
-            }
-
-            static RejectOrder rejectOrder() {
-                return REJECT_ORDER;
-            }
-
-            static AssignRegistrantDetails assignRegistrantDetails() {
-                return ASSIGN_REGISTRANT_DETAILS;
-            }
-        }
-
-        private static class Event {
-
-            static final OrderPlaced ORDER_PLACED = OrderPlaced.newBuilder()
-                    .setOrderId(ORDER_ID)
-                    .setConferenceId(CONFERENCE_ID)
-                    .addAllSeat(SEATS)
-                    .build();
-
-            private static final OrderConfirmed ORDER_CONFIRMED = OrderConfirmed.newBuilder().setOrderId(ORDER_ID).build();
-
-            static OrderPlaced orderPlaced() {
-                return ORDER_PLACED;
-            }
-
-            static OrderUpdated orderUpdated() {
-                final OrderUpdated.Builder result = OrderUpdated.newBuilder()
-                        .setOrderId(ORDER_ID)
-                        .addSeat(newSeatQuantity(16))
-                        .addSeat(newSeatQuantity(32));
-                return result.build();
-            }
-
-            static OrderPartiallyReserved orderPartiallyReserved() {
-                final OrderPartiallyReserved.Builder result = OrderPartiallyReserved.newBuilder()
-                        .setOrderId(ORDER_ID)
-                        .addSeat(newSeatQuantity(64))
-                        .addSeat(newSeatQuantity(128));
-                return result.build();
-            }
-
-            static OrderReservationCompleted orderReservationCompleted() {
-                final OrderReservationCompleted.Builder result = OrderReservationCompleted.newBuilder()
-                        .setOrderId(ORDER_ID)
-                        .addSeat(newSeatQuantity(256))
-                        .addSeat(newSeatQuantity(512));
-                return result.build();
-            }
-
-            static OrderConfirmed orderConfirmed() {
-                return ORDER_CONFIRMED;
-            }
-        }
-    }
-
-    private static class PricingServiceStub implements OrderPricingService {
+    static class PricingServiceStub implements OrderPricingService {
 
         private static final Money TOTAL_PRICE = newMoney(100, USD);
         private static final SeatOrderLine ORDER_LINE = SeatOrderLine.newBuilder()
